Allow checkpointer to override logic that creates the next version for each channel
