+++ langgraph/channels/base.py
@@ -1,12 +1,10 @@
 from abc import ABC, abstractmethod
 from contextlib import asynccontextmanager, contextmanager
-from datetime import datetime, timezone
 from typing import (
     Any,
     AsyncGenerator,
     Generator,
     Generic,
-    Mapping,
     Optional,
     Sequence,
     TypeVar,
@@ -14,8 +12,6 @@ from typing import (
 
 from typing_extensions import Self
 
-from langgraph.checkpoint.base import Checkpoint
-from langgraph.checkpoint.id import uuid6
 from langgraph.errors import EmptyChannelError, InvalidUpdateError
 
 Value = TypeVar("Value")
@@ -83,68 +79,8 @@ class BaseChannel(Generic[Value, Update, C], ABC):
         pass
 
 
-@contextmanager
-def ChannelsManager(
-    channels: Mapping[str, BaseChannel],
-    checkpoint: Checkpoint,
-) -> Generator[Mapping[str, BaseChannel], None, None]:
-    """Manage channels for the lifetime of a Pregel invocation (multiple steps)."""
-    # TODO use https://docs.python.org/3/library/contextlib.html#contextlib.ExitStack
-    empty = {
-        k: v.from_checkpoint(checkpoint["channel_values"].get(k))
-        for k, v in channels.items()
-    }
-    try:
-        yield {k: v.__enter__() for k, v in empty.items()}
-    finally:
-        for v in empty.values():
-            v.__exit__(None, None, None)
-
-
-@asynccontextmanager
-async def AsyncChannelsManager(
-    channels: Mapping[str, BaseChannel],
-    checkpoint: Checkpoint,
-) -> AsyncGenerator[Mapping[str, BaseChannel], None]:
-    """Manage channels for the lifetime of a Pregel invocation (multiple steps)."""
-    empty = {
-        k: v.afrom_checkpoint(checkpoint["channel_values"].get(k))
-        for k, v in channels.items()
-    }
-    try:
-        yield {k: await v.__aenter__() for k, v in empty.items()}
-    finally:
-        for v in empty.values():
-            await v.__aexit__(None, None, None)
-
-
-def create_checkpoint(
-    checkpoint: Checkpoint, channels: Mapping[str, BaseChannel], step: int
-) -> Checkpoint:
-    """Create a checkpoint for the given channels."""
-    ts = datetime.now(timezone.utc).isoformat()
-    values: dict[str, Any] = {}
-    for k, v in channels.items():
-        try:
-            values[k] = v.checkpoint()
-        except EmptyChannelError:
-            pass
-    return Checkpoint(
-        v=1,
-        ts=ts,
-        id=str(uuid6(clock_seq=step)),
-        channel_values=values,
-        channel_versions=checkpoint["channel_versions"],
-        versions_seen=checkpoint["versions_seen"],
-        pending_sends=checkpoint.get("pending_sends", []),
-    )
-
-
 __all__ = [
     "BaseChannel",
-    "ChannelsManager",
-    "AsyncChannelsManager",
-    "create_checkpoint",
     "EmptyChannelError",
     "InvalidUpdateError",
 ]
