+++ tests/checkpoint/test_aiosqlite.py
@@ -1,7 +1,7 @@
 import pytest
 from langchain_core.runnables import RunnableConfig
 
-from langgraph.channels.base import create_checkpoint
+from langgraph.channels.manager import create_checkpoint
 from langgraph.checkpoint.aiosqlite import AsyncSqliteSaver
 from langgraph.checkpoint.base import Checkpoint, CheckpointMetadata, empty_checkpoint
